[chore]#101 패키지 정리
