+++ app/src/main/AndroidManifest.xml
@@ -39,7 +39,7 @@
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".presentation.schedule.search.view.SearchActivity"
+            android:name=".presentation.mission.search.view.SearchActivity"
             android:exported="false"
             android:screenOrientation="portrait" />
         <service
@@ -55,7 +55,7 @@
             android:exported="true"
             android:screenOrientation="portrait"/>
         <activity
-            android:name=".presentation.schedule.addition.view.AdditionActivity"
+            android:name=".presentation.mission.addition.view.AdditionActivity"
             android:exported="false"
             android:screenOrientation="portrait"
             android:windowSoftInputMode="adjustPan" />
