--- app/src/main/java/kr/co/nottodo/data/remote/model/Test.kt
@@ -1,4 +1,0 @@
-package kr.co.nottodo.data.remote.model
-
-class Test {
-}
