+++ app/src/main/java/kr/co/nottodo/presentation/MainActivity.kt
@@ -12,7 +12,6 @@ import kr.co.nottodo.databinding.ActivityMainBinding
 import kr.co.nottodo.presentation.achievement.view.AchievementFragment
 import kr.co.nottodo.presentation.home.HomeFragment
 import kr.co.nottodo.presentation.mypage.view.MyPageFragment
-import kr.co.nottodo.presentation.schedule.addition.view.AdditionActivity
 
 import kr.co.nottodo.presentation.toplevel.recommendation.recommendationfragment.RecommendationFragment
