+++ app/src/main/java/kr/co/nottodo/presentation/addsituation/viewmodel/AddSituationViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.*
 import kotlinx.coroutines.launch
 import kr.co.nottodo.data.remote.api.ServicePool
 import kr.co.nottodo.data.remote.model.ResponseAddSituationDto
-import kr.co.nottodo.presentation.schedule.addition.view.AdditionActivity.Companion.BLANK
+import kr.co.nottodo.presentation.mission.addition.view.AdditionActivity.Companion.BLANK
 
 class AddSituationViewModel : ViewModel() {
     private val addSituationService by lazy { ServicePool.addSituationService }
