+++ app/src/main/java/kr/co/nottodo/presentation/recommendation/recommendationfragment/RecommendationViewHolder.kt
@@ -1,0 +1,27 @@
+package kr.co.nottodo.presentation.toplevel.recommendation.recommendationfragment
+
+import androidx.recyclerview.widget.RecyclerView
+import com.bumptech.glide.Glide
+import kr.co.nottodo.databinding.ItemRecommendBinding
+import kr.co.nottodo.data.remote.model.ResponseRecommendationCategorySituationDto
+
+class RecommendationViewHolder(
+    val binding: ItemRecommendBinding
+) : RecyclerView.ViewHolder(binding.root){
+
+    private lateinit var recommendationItem: ResponseRecommendationCategorySituationDto.CategorySituation
+
+    fun interface RecommendationClickDelegate {
+        fun onClickRecommendation(
+            recommendItem: ResponseRecommendationCategorySituationDto.CategorySituation
+        )
+    }
+
+    fun onBind(data: ResponseRecommendationCategorySituationDto.CategorySituation) {
+        recommendationItem = data
+        binding.tvRecommendTitle.text = data.name
+        Glide.with(binding.root.context)
+            .load(data.image)
+            .into(binding.ivRecommend)
+    }
+}
