--- app/src/main/java/kr/co/nottodo/presentation/toplevel/recommendation/data/recommendationlistdata/RecommendationChildData.kt
@@ -1,6 +1,0 @@
-package kr.co.nottodo.presentation.toplevel.recommendation.data.recommendationlistdata
-
-data class RecommendationChildData(
-    val id: Int,
-    val title: String
-)
