--- app/src/main/java/kr/co/nottodo/presentation/toplevel/recommendation/data/responsedto/ResponseBase.kt
@@ -1,11 +1,0 @@
-package kr.co.nottodo.presentation.toplevel.recommendation.data.responsedto
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ResponseBase<T>(
-    val status: Int,
-    val success: Boolean,
-    val message: String,
-    val data: T,
-)
