--- app/src/main/java/kr/co/nottodo/presentation/toplevel/recommendation/viewmodel/AddSituationViewModel.kt
@@ -1,4 +1,0 @@
-package kr.co.nottodo.presentation.toplevel.recommendation.viewmodel
-
-class AddSituationViewModel {
-}
